requirements.txt

# Automatically generated by https://github.com/damnever/pigar.

alembic==1.10.2
Flask==2.2.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-Reuploaded==1.3.0
marshmallow==3.19.0
marshmallow_sqlalchemy==0.29.0
Pillow==9.5.0
PyJWT==2.6.0
SQLAlchemy==2.0.7
Werkzeug==2.2.3

